Lv2 Post 엔티티의 id 타입 Integer -> Long 변경 Post update 와 delete 로직 변경( post에 해당하는 userId와 요청한 userId가 일치하는지 확인)

+++ src/main/java/com/example/memo/repository/PostJpaRepository.java
@@ -6,10 +6,10 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostJpaRepository extends JpaRepository<Post,Integer> {
+public interface PostJpaRepository extends JpaRepository<Post,Long> {
 
     List<Post> findAllByOrderByModifiedAtDesc();
 
-    Optional<Post> findByIdAndUser(Integer id, User user);
+    Optional<Post> findByIdAndUser(Long id, User user);
 
 }
